+++ Assets/Code/Runtime/Logic/WeaponSystem/IBullet.cs
@@ -5,6 +5,6 @@ namespace Code.Runtime.Logic.WeaponSystem
     public interface IBullet
     {
         void Initialize(int damage);
-        void Move(Vector2 direction, int force);
+        void Move(Vector2 direction, float force);
     }
 }
